+++ databricks-sdk-java/src/main/java/com/databricks/sdk/service/jobs/ListRunsRequest.java
@@ -28,7 +28,11 @@ public class ListRunsRequest {
   @QueryParam("completed_only")
   private Boolean completedOnly;
 
-  /** Whether to include task and cluster details in the response. */
+  /**
+   * Whether to include task and cluster details in the response. Note that in API 2.2, only the
+   * first 100 elements will be shown. Use :method:jobs/getrun to paginate through all tasks and
+   * clusters.
+   */
   @JsonIgnore
   @QueryParam("expand_tasks")
   private Boolean expandTasks;