+++ databricks-sdk-java/src/main/java/com/databricks/sdk/service/pipelines/RestartWindow.java
@@ -15,7 +15,7 @@ public class RestartWindow {
    * start_hour). If not specified all days of the week will be used.
    */
   @JsonProperty("days_of_week")
-  private Collection<RestartWindowDaysOfWeek> daysOfWeek;
+  private Collection<DayOfWeek> daysOfWeek;
 
   /**
    * An integer between 0 and 23 denoting the start hour for the restart window in the 24-hour day.
@@ -32,12 +32,12 @@ public class RestartWindow {
   @JsonProperty("time_zone_id")
   private String timeZoneId;
 
-  public RestartWindow setDaysOfWeek(Collection<RestartWindowDaysOfWeek> daysOfWeek) {
+  public RestartWindow setDaysOfWeek(Collection<DayOfWeek> daysOfWeek) {
     this.daysOfWeek = daysOfWeek;
     return this;
   }
 
-  public Collection<RestartWindowDaysOfWeek> getDaysOfWeek() {
+  public Collection<DayOfWeek> getDaysOfWeek() {
     return daysOfWeek;
   }